me03#29261 — Order Line Split: move listener to .impl. subpackage to match Services.get convention
